rename pub-h1 and pub-wrapper into pub-page and pub-page-title
