+++ publiq-ui/pub-page.stories.mdx
@@ -1,13 +1,13 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs/blocks';
-import PubH1 from './pub-h1';
+import PubPage from './pub-page.vue';
 
-<Meta title="Components/pub-h1" component={PubH1} />
+<Meta title="Components/pub-page" component={PubPage} />
 
 export const Template = (args, { argTypes }) => {
   return {
     props: Object.keys(argTypes),
-    components: { PubH1 },
-    template: '<pub-h1 v-bind="$props">Heading 1</pub-h1>',
+    components: { PubPage },
+    template: '<pub-page v-bind="$props">Page content</pub-page>',
   };
 };
 
@@ -18,7 +18,7 @@ export const Template = (args, { argTypes }) => {
 </Canvas>
 
 ```javascript
-<pub-h1>Heading 1</pub-h1>
+<pub-page>Page content</pub-page>
 ```
 
 <ArgsTable story="Default" />
